initial poc for dynamic size memory accounts

+++ program/src/instruction/init_memory.rs
@@ -82,8 +82,8 @@ pub fn process_init_memory(accounts: &[AccountInfo<'_>], data: &[u8]) -> Program
     memory.name = args.name;
     memory.vm = vm_info.key.clone();
     memory.bump = args.vm_memory_bump;
-    memory.num_accounts = args.num_accounts;
-    memory.account_size = args.account_size;
+    memory.num_accounts = args.num_accounts as u32;
+    memory.account_size = args.account_size as u16;
 
     vm.advance_poh(CodeInstruction::InitMemoryIx, accounts, data);
